requirements.txt
aiogram==3.*
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.3
requests==2.31.0 